emu: add stack sampler to emulated CPU
Introduce an optional stack sampler to the CPU, which traces JAL and
JALR instructions to maintain a shadow call stack, and record a stack
sample periodically under the given sample rate.

+++ sw-emulator/lib/cpu/src/instr/test_macros.rs
@@ -467,8 +467,10 @@ mod test {
             use caliptra_emu_bus::{Clock, DynamicBus, Ram, Rom};
             use $crate::cpu::Cpu;
 
-            let text_range = $text_addr..=u32::try_from($text_addr + $text.len() - 1).unwrap();
-            let data_range = $data_addr..=u32::try_from($data_addr + $data.len() - 1).unwrap();
+            let text_range =
+                $text_addr..=u32::try_from($text_addr + $text.len() as u32 - 1).unwrap();
+            let data_range =
+                $data_addr..=u32::try_from($data_addr + $data.len() as u32 - 1).unwrap();
 
             let mut cpu = Cpu::new(DynamicBus::new(), Clock::new());
             let rom = Rom::new($text.clone());
